final fixes

+++ server/src/main/java/bgu/spl/net/srv/ConnectionsImpl.java
@@ -1,4 +1,5 @@
 package bgu.spl.net.srv;
+
 import java.util.ArrayList;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.Collections;
@@ -7,14 +8,14 @@ import java.util.concurrent.atomic.*;
 
 import bgu.spl.net.impl.stomp.StompFrame;
 
-public class ConnectionsImpl <T> implements Connections <T> {
-    private ConcurrentHashMap<Integer, ConnectionHandler<T>> connections; 
+public class ConnectionsImpl<T> implements Connections<T> {
+    private ConcurrentHashMap<Integer, ConnectionHandler<T>> connections;
     private ConcurrentHashMap<String, ArrayList<Integer>> topics;
     private ConcurrentHashMap<String, User> active_logins;
     private ConcurrentHashMap<String, User> inactive_logins;
     private AtomicInteger currentMessageID;
 
-    public ConnectionsImpl(){
+    public ConnectionsImpl() {
         connections = new ConcurrentHashMap<>();
         topics = new ConcurrentHashMap<>();
         active_logins = new ConcurrentHashMap<>();
@@ -22,69 +23,73 @@ public class ConnectionsImpl <T> implements Connections <T> {
         currentMessageID = new AtomicInteger(0);
     }
 
-    public void addConnection(int connectionId, ConnectionHandler<T> handler){
+    public void addConnection(int connectionId, ConnectionHandler<T> handler) {
         connections.put(connectionId, handler);
     }
 
-    public int getAndIncrementMessageId(){
+    public int getAndIncrementMessageId() {
         return currentMessageID.getAndIncrement();
     }
-    public boolean subscribeToTopic(String topic, int subscriptionId, int connectionId){
+
+    public boolean subscribeToTopic(String topic, int subscriptionId, int connectionId) {
         // check if user is already subscribed to topic and if so return false
-        if (active_logins.get(connections.get(connectionId).getUser().getUsername()).getSubscribedTopics().containsValue(topic)){
+        if (active_logins.get(connections.get(connectionId).getUser().getUsername()).getSubscribedTopics()
+                .containsValue(topic)) {
             return false;
         }
 
-        if(!topics.containsKey(topic)){
+        if (!topics.containsKey(topic)) {
             topics.put(topic, new ArrayList<>());
         }
         topics.get(topic).add(connectionId);
-        active_logins.get(connections.get(connectionId).getUser().getUsername()).subscribeToTopic(topic, subscriptionId);
+        active_logins.get(connections.get(connectionId).getUser().getUsername()).subscribeToTopic(topic,
+                subscriptionId);
         return true;
     }
 
-    public boolean unsubscribeFromTopic(int subscriptionId, int connectionId){
+    public boolean unsubscribeFromTopic(int subscriptionId, int connectionId) {
         User user = connections.get(connectionId).getUser();
         String topic = user.getSubscribedTopics().get(subscriptionId);
         // check if user is subscribed to topic and if not return false
-        if (topic == null){
+        if (topic == null) {
             return false;
         }
 
-        topics.get(topic).remove(connectionId);
+        // topics.get(topic).remove(connectionId);
+        topics.get(topic).remove((Integer) connectionId);
+
         user.removeSubscribedTopic(subscriptionId);
         return true;
     }
 
-
-    public String login(String username, String password, int connectionId){
+    public String login(String username, String password, int connectionId) {
         // check if user is already logged in
-        if(active_logins.containsKey(username)){
+        if (active_logins.containsKey(username)) {
             return "already logged in";
         }
 
         // check if user is new
         User user = inactive_logins.get(username);
-        if(user == null){
+        if (user == null) {
             loginNewUser(username, password, connectionId);
             return "connected";
         }
 
         // check if password is correct for existing user
-        if(user.getPassword().equals(password)){
+        if (user.getPassword().equals(password)) {
             loginExistingUser(username, connectionId);
             return "connected";
         }
         return username;
     }
 
-    public void loginNewUser(String username, String password, int connectionId){
+    public void loginNewUser(String username, String password, int connectionId) {
         User user = new User(username, password, connectionId);
         active_logins.put(username, user);
         connections.get(connectionId).setUser(user);
     }
 
-    public void loginExistingUser(String username, int connectionId){
+    public void loginExistingUser(String username, int connectionId) {
         User user = inactive_logins.get(username);
         active_logins.put(username, user);
         inactive_logins.remove(username);
@@ -104,7 +109,8 @@ public class ConnectionsImpl <T> implements Connections <T> {
         for (Integer connectionId : topics.get(channel)) {
             user = connections.get(connectionId).getUser();
             subscriptionId = user.getSubscriptionIdForTopic(channel);
-            send(connectionId, connections.get(connectionId).getProtocol().addSubscriptionIdToMessage(msg, subscriptionId));
+            send(connectionId,
+                    connections.get(connectionId).getProtocol().addSubscriptionIdToMessage(msg, subscriptionId));
         }
     }
 
@@ -115,13 +121,13 @@ public class ConnectionsImpl <T> implements Connections <T> {
         ConcurrentHashMap<Integer, String> subscribedTopics = user.getSubscribedTopics();
 
         // remove user from all topics
-        for(int subscriptionId : Collections.list(subscribedTopics.keys())){
+        for (int subscriptionId : Collections.list(subscribedTopics.keys())) {
             String topic = subscribedTopics.get(subscriptionId);
-            topics.get(topic).remove(connectionId);
+            topics.get(topic).remove((Integer) connectionId);
         }
 
         // remove user from active logins, connections and unsubscribe from all topics
-        connections.remove(connectionId);
+        connections.remove((Integer) connectionId);
         user.unsuscribeFromAllTopics();
         active_logins.remove(user.getUsername());
         inactive_logins.put(user.getUsername(), user);
@@ -129,7 +135,8 @@ public class ConnectionsImpl <T> implements Connections <T> {
 
     @Override
     public boolean isUserSubscribedToTopic(int connectionId, String topic) {
-        return active_logins.get(connections.get(connectionId).getUser().getUsername()).getSubscribedTopics().containsValue(topic);
+        return active_logins.get(connections.get(connectionId).getUser().getUsername()).getSubscribedTopics()
+                .containsValue(topic);
     }
 
 }
